+++ src/index.js
@@ -1,3 +1,5 @@
 import Countdown360 from './countdown360'
 
 export default Countdown360
+export * from './timeFormatters'
+export * from './unitFormatters'